Fixed text positioning (mostly)
The text (name and title) now sticks to the photo in the way I want.
This had to be done with JS in the end, rather than purely with CSS.

The distance between the two lines is still static, and so looks bad
depending on the screen size. I need to change it to be dynamically
sized like the positioning of the whole TypedTitle component.